requirements.txt
pypdf
pytest-playwright
pyyaml
tqdm